+++ components/settings_ex/settings.c
@@ -49,6 +49,7 @@ esp_err_t settings_init(void)
         s_settings.bluetooth.enable         = false;
         s_settings.bluetooth.show_screen    = true;
         s_settings.bluetooth.volume         = 15;
+        s_settings.bluetooth.auto_switch    = true;
         strncpy(s_settings.ntp.server1, "pool.ntp.org",                sizeof(s_settings.ntp.server1) - 1);
         strncpy(s_settings.ntp.server2, "time.cloudflare.com",         sizeof(s_settings.ntp.server2) - 1);
         strncpy(s_settings.ntp.tz,      "CET-1CEST,M3.5.0,M10.5.0/3", sizeof(s_settings.ntp.tz)      - 1);
@@ -226,6 +227,10 @@ static esp_err_t load_from_file(void)
         cJSON *bt_vol = cJSON_GetObjectItem(bluetooth, "volume");
         if (cJSON_IsNumber(bt_vol)) {
             s_settings.bluetooth.volume = bt_vol->valueint;
+        }
+        cJSON *bt_auto = cJSON_GetObjectItem(bluetooth, "auto_switch");
+        if (cJSON_IsBool(bt_auto)) {
+            s_settings.bluetooth.auto_switch = cJSON_IsTrue(bt_auto);
         }
     }
 
@@ -410,6 +415,7 @@ static esp_err_t save_to_file(void)
     cJSON_AddBoolToObject(bluetooth,   "enable", s_settings.bluetooth.enable);
     cJSON_AddBoolToObject(bluetooth,   "show_screen", s_settings.bluetooth.show_screen);
     cJSON_AddNumberToObject(bluetooth, "volume", s_settings.bluetooth.volume);
+    cJSON_AddBoolToObject(bluetooth,   "auto_switch", s_settings.bluetooth.auto_switch);
     cJSON_AddItemToObject(json, "bluetooth", bluetooth);
 
     // ntp
@@ -502,6 +508,7 @@ void settings_apply(void)
         .has_bt_enable          = true, .bt_enable = s_settings.bluetooth.enable,
         .has_bt_show_screen     = true, .bt_show_screen = s_settings.bluetooth.show_screen,
         .has_bt_volume          = true, .bt_volume = s_settings.bluetooth.volume,
+        .has_bt_auto_switch     = true, .bt_auto_switch = s_settings.bluetooth.auto_switch,
         .has_display_brightness = true, .display_brightness = s_settings.display.brightness,
         .has_theme              = true, .theme     = s_settings.display.theme,
         .has_bg_gradient        = true, .bg_gradient = s_settings.display.bg_gradient,
@@ -619,6 +626,15 @@ void settings_set_bt_show_screen(bool show)
     if(s_settings.bluetooth.show_screen != show) {
         s_settings.bluetooth.show_screen = show;
         app_state_update(&(app_state_patch_t){ .has_bt_show_screen = true, .bt_show_screen = show });
+        save_to_file();
+    }
+}
+
+void settings_set_bt_auto_switch(bool enable)
+{
+    if(s_settings.bluetooth.auto_switch != enable) {
+        s_settings.bluetooth.auto_switch = enable;
+        app_state_update(&(app_state_patch_t){ .has_bt_auto_switch = true, .bt_auto_switch = enable });
         save_to_file();
     }
 }